Configurable HTTP host address
Before commit 223dc8dd91a5 where both the web server and the mail server
listingen on 0.0.0.0. also known as the any address.
It did mean that the web server was always exposed to the whole world.

Commit 223dc8dd91a5 made it possible to (optional) configure the bind
address, with default to 127.0.0.1.  Both, web and mail, to same value.
Resulting in having the web server listening only on localhost, ment
also the mail server listening ONLY on 127.0.0.1.

This change binds mail server hardcoded to 0.0.0.0 and web server
is configurable with environment variable HTTP_HOST, defaulting to
127.0.0.1

Signed-off-by: Geert Stappers <stappers@stappers.it>