Have version information available
It is considered a good thing knowing to whom are you talking. For
making that possible is the HTTP API extended with end point named
'version' and tells SMTP server tells after 'MAIL FROM' which version
it is.
Both "servers" get the version information from the Cargo.toml of the
backend.
In the source code are now a NOTE for telling that the response to the
SMTP HELO is as just returning a '250 further info'.
And there is a TODO that documents that the returned JSON is not JSON.